requirements.txt
rev-ai==2.17.1
openai
pyttsx3
tkinter
azure-cognitiveservices-speech==1.29.0
mtranslate
googlemaps
pyaudio